+++ src/сontroller/Main.java
@@ -10,18 +10,18 @@ public class Main {
         TaskManager taskManager = managers.getDefault();
         HistoryManager historyManager = Managers.getDefaultHistory();
 
-        int firstEpic = taskManager.createEpic(new Epic("Эпик 1", "Описание 1", Status.NEW));
-        int firstSubTask = taskManager.createSubTask(new SubTask("Подзадача 1",
+        long firstEpic = taskManager.createEpic(new Epic("Эпик 1", "Описание 1", Status.NEW));
+        long firstSubTask = taskManager.createSubTask(new SubTask("Подзадача 1",
                "Описание", Status.NEW, firstEpic));
-        int secondSubTask = taskManager.createSubTask(new SubTask("Подзадача 2",
+        long secondSubTask = taskManager.createSubTask(new SubTask("Подзадача 2",
                "Описание", Status.DONE, firstEpic));
 
-        int secondEpic = taskManager.createEpic(new Epic("Эпик 2", "Описание 2", Status.IN_PROGRESS));
-        int firstSubTaskSecondEpic = taskManager.createSubTask(new SubTask("Подзадача 1 второго эпика",
+        long secondEpic = taskManager.createEpic(new Epic("Эпик 2", "Описание 2", Status.IN_PROGRESS));
+        long firstSubTaskSecondEpic = taskManager.createSubTask(new SubTask("Подзадача 1 второго эпика",
                 "Описание", Status.DONE, secondEpic));
 
-        int firstTask = taskManager.createTask(new Task("Задача 1", "Описание", Status.IN_PROGRESS));
-        int secondTask = taskManager.createTask(new Task("Задача 2", "Описание", Status.DONE));
+        long firstTask = taskManager.createTask(new Task("Задача 1", "Описание", Status.IN_PROGRESS));
+        long secondTask = taskManager.createTask(new Task("Задача 2", "Описание", Status.DONE));
 
         taskManager.getByID(firstEpic);
         taskManager.getByID(firstEpic);
@@ -51,8 +51,6 @@ public class Main {
         System.out.println(taskManager.getTaskList());
         System.out.println("История" + historyManager.getHistory());
 
-        //taskManager.deleteByID(firstEpic);
-        //taskManager.deleteByID(firstTask);
         taskManager.deleteAllEpic();
         System.out.println(taskManager.getEpicList());
         System.out.println("История" + historyManager.getHistory());
