+++ src/сontroller/TaskManager.java
@@ -6,19 +6,19 @@ import tasks.Task;
 import java.util.List;
 
 public interface TaskManager {
-    int createTask(Task newTask);
-    int createEpic(Epic newEpic);
-    int createSubTask(SubTask newSubTask);
-    void updateTask (int id, Task newTask);
-    void updateEpic(int id, Epic newEpic);
-    void updateSubTask(int id, SubTask newSubTask);
+    long createTask(Task newTask);
+    long createEpic(Epic newEpic);
+    long createSubTask(SubTask newSubTask);
+    void updateTask (long id, Task newTask);
+    void updateEpic(long id, Epic newEpic);
+    void updateSubTask(long id, SubTask newSubTask);
     List<Task> getTaskList();
     List<Epic> getEpicList();
     List<SubTask> getSubTaskList();
     void deleteAllTask();
     void deleteAllEpic();
     void deleteAllSubTask();
-    void deleteByID(int id);
-    Task getByID(int id);
-    List<SubTask> getEpicSubTasks(int epicID);
+    void deleteByID(long id);
+    Task getByID(long id);
+    List<SubTask> getEpicSubTasks(long epicID);
 }
